Correct remainder of crate::configuration::*
